Refactor: moved todo toggle logic to ToDoItem
