+++ src/components/EventInfo.js
@@ -6,7 +6,7 @@ import { useParams, Link } from 'react-router-dom';
 
 export default function EventInfo(props) {
   const { eventId } = useParams();
-  const { eventData, setEventData } = useContext(EventsContext);
+  const { eventData } = useContext(EventsContext);
   const event = eventData.find((e) => e.id === eventId);
 
   if (event === undefined) {
@@ -14,26 +14,6 @@ export default function EventInfo(props) {
   }
 
   const todos = event.todos;
-
-  function handleToDoDoneClick(todoClicked) {
-    const doneTodos = todos.map((todo) => {
-      if (todo.id === todoClicked.id) {
-        return { ...todo, status: !todo.status };
-      } else {
-        return todo;
-      }
-    });
-    // updates EventsContext
-    setEventData(
-      eventData.map((e) => {
-        if (event.id === e.id) {
-          return { ...e, todos: doneTodos };
-        } else {
-          return e;
-        }
-      })
-    );
-  }
 
   return (
     <li className='list-none'>
@@ -65,12 +45,7 @@ export default function EventInfo(props) {
       </div>
       <ul className='bg-red-400 text-black rounded-3xl flex flex-col items-center justify-center shadow-2xl py-6 px-6 max-w-full'>
         {todos.map((todo, index) => (
-          <ToDoItem
-            key={index}
-            todo={todo}
-            onDoneClick={handleToDoDoneClick}
-            prio='hidden'
-          />
+          <ToDoItem key={index} eventId={eventId} todo={todo} prio='hidden' />
         ))}
       </ul>
     </li>
