+++ src/components/ToDoItem.js
@@ -1,10 +1,30 @@
-export default function ToDoItem(props) {
-  const { todo } = props;
+import { useContext } from 'react';
+import { EventsContext } from '../contexts/EventsContext';
 
-  function handleDoneClick() {
-    console.log('ToDo done clicked');
-    props.onDoneClick(todo);
-  }
+const ToDoItem = (props) => {
+  const { todo, eventId } = props;
+  const { eventData, setEventData } = useContext(EventsContext);
+
+  const toggleTodo = (todoClicked) => {
+    const todos = eventData.find((e) => e.id === eventId).todos;
+    const doneTodos = todos.map((todo) => {
+      if (todo.id === todoClicked.id) {
+        return { ...todo, status: !todo.status };
+      } else {
+        return todo;
+      }
+    });
+    // updates EventsContext
+    setEventData(
+      eventData.map((e) => {
+        if (eventId === e.id) {
+          return { ...e, todos: doneTodos };
+        } else {
+          return e;
+        }
+      })
+    );
+  };
 
   return (
     <li className='flex flex-row items-center mb-8 w-full'>
@@ -21,7 +41,7 @@ export default function ToDoItem(props) {
         type='checkbox'
         id={todo.id}
         className='form-checkbox rounded-full w-12 h-12 ml-2 text-blue-400 focus:ring-0'
-        onChange={handleDoneClick}
+        onChange={() => toggleTodo(todo)}
         checked={todo.status}
       />
       <label
@@ -33,4 +53,5 @@ export default function ToDoItem(props) {
       </label>
     </li>
   );
-}
+};
+export default ToDoItem;
